requirements.txt
python-dateutil
cryptography
pydantic
PyYAML
pytest
pytest-cov